Handle uploading source versions of binaries when needed.
When a change involving binary files is uploaded that requires a parent
diff, we need to upload both the modified and source versions of the
binary file (since the server can't fetch the source version from the
repository). This change handles that, with the new `source_file`
parameter to the diff file attachments API.

Testing Done:
Uploaded changes to binary files that required a parent diff. Saw the
correct version of the source file.

Reviewed at https://reviews.reviewboard.org/r/13608/