Implement binary files support for Mercurial.
This change implements the get_file_content and get_file_size methods
for the Mercurial client backend.

Additionally, this changes the log methods to always use `{node}`
instead of `{node|short}`. We had a mix (individual commits were using
the full node SHA, while the cumulative diff was using the short form),
which was causing issues with file uploads--we'd upload a file for the
commit version with the full SHA revision, and then not find it because
the cumulative diff was using the short SHA.

Testing Done:
- Ran unit tests.
- Uploaded a variety of revisions including binary files. Saw that
  everything worked correctly.

Reviewed at https://reviews.reviewboard.org/r/13610/