Make the user agent accessible from the RBClient class.
This makes the user agent string used for clients accessible from `RBClient`
as a property. We also weren't passing on the user agent from `SyncTransport`
to the `ReviewBoardServer`, so if a custom agent was set on the
`SyncTransport`, the `ReviewBoardServer` would still use the default agent.
This change fixes that.

Testing Done:
- Ran unit tests.
- Used in an upcoming change where I access the user agent from the
  `RBClient` in a command.

Reviewed at https://reviews.reviewboard.org/r/13340/