Make sure that calls to make_tempfile() use keyword arguments.
I recently used `rbt patch` and saw that it was broken. The problem was that
it didn't pass keyword arguments with `make_tempfile()`, which has
keyword-only arguments. I audited the rest of the code to see if we had this
problem anywhere else, but this was the only spot.

Testing Done:
- Ran unit tests.
- Used `rbt patch` and saw that it works now.

Reviewed at https://reviews.reviewboard.org/r/13701/