Add a flag for enabling web-based login and update docs.
This adds a `.reviewboardrc` flag for enabling web-based login, which we leave
off by default. The reason why we leave it off by default is because we don't
currently support web-based login when authenticating RBTools on a remote
system, so it's better to let users opt-in to the feature rather than have
it on by default.

This also includes a couple of other small changes:
1. Updates the docs related to authentication.
2. Fixes a bug where we used the wrong `config_key` for the open browser
   option in `rbt login`.
3. Allows web-based login to be used when authenticating users through
   commands other than `rbt login`.

Testing Done:
- Tested `rbt login`, saw that web-based login was turned off by default and
  could use it by passing the `--web` option or setting `WEB_LOGIN` in
  `.reviewboardrc`
- Tested using `rbt status` without being logged in, saw that it would bring
  me to web-based login if it was enabled, and prompted for username and
  password if it wasn't enabled.

Reviewed at https://reviews.reviewboard.org/r/13702/