Don't try to upload source for new binary files added in child diffs.
The fix to upload source versions of binary files when posting changes
that include parent diffs ended up causing warnings when attempting to
post a change that added a new binary file.

This change adds a check so we don't try to upload the source version of
a binary file if the source revision is PRE-CREATION.

Testing Done:
Created a branch on top of another branch and made a commit that added a
new binary file. Posted the diff between the two branches and saw that I
did not see any warnings about trying to get the file content at the
PRE-CREATION revision.

Reviewed at https://reviews.reviewboard.org/r/13771/