Fix uploading binary parent diffs.
When uploading changes to a binary file in a diff, we were referencing a
local variable for the diff file attachments resource that didn't exist.
Instead, the instance lives on the command instance.

This is a simple change that references that instance, fixing binary
diff upload.

Testing Done:
Successfully posted a change to some binary image files.

Reviewed at https://reviews.reviewboard.org/r/13886/