Make sure all commands fall back on normal repository detection code if not given a repository type.
In the past we've ran into errors where RBTools commands weren't properly
trying to find the right repository to use, and weren't outputting a nice
error message when it failed to find repositories. We had a [fix](/r/13101/) for this,
but we missed one are where we also needed to apply the fix. This change adds
the fix for that area, and makes sure we've fixed this everywhere.

Testing Done:
- Ran `rbt login` in a random directory with no configured repository, saw
  a proper error message.
- Ran unit tests.
- Audited the rest of the code for this problem.

Reviewed at https://reviews.reviewboard.org/r/13933/