Handle parent diff binary file duplicates.
When iterating on a branch involving binary files, it was possible to
hit a case where we would try to upload the source file associated with
a parent diff and hit a conflict. The API returns a DUPLICATE_ITEM error
because we already have a file with that revision, and RBTools would
then crash.

Because of the way file attachment comments work on diffs, it's fine for
us to share the source file attachment for the parent diff across
multiple different diff revisions. We can therefore ignore the duplicate
error.

Testing Done:
Followed the reproduction steps for the bug and saw that before this
change `rbt post` would crash, and after everything worked fine.
Verified that the posted diffs worked as expected.

Reviewed at https://reviews.reviewboard.org/r/14052/