Fix problems with `rbt status-update set` and documentation.
The documentation for creating reviews with `rbt status-update set`
indicated that body_top and body_bottom content should be set inside of
a `review` key in the JSON file, but the actual implementation was using
`reviews`. `review` is much more appropriate, so I've changed the
implementation to match the documentation.

This also adds `issue_opened` as a field for general comments in the
documentation. Because of how we create comments, any field that's valid
for the comment resource will work, but it was confusing to list
`issue_opened` under diff comments but not general comments.

Testing Done:
Posted a review using `rbt status-update set` that included body top and
body bottom content, and saw that that content was properly reflected in
the rendered status update.

Reviewed at https://reviews.reviewboard.org/r/14056/