Suppress warnings when testing RBTools version help output.
We had a unit test regression in RBTools 2 that was caused by
`pkg_resources` emitting a deprecation warning on import. This resulted
in the version output not being quite what we expected.

While RBTools 2 is no longer under development, this same issue is
likely going to crop up in the future due to other imports, and it's
best to take care of it now.

We now ignore warnings when we execute `rbt` for version testing
purposes, using `python -W ignore`.

Testing Done:
Unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/14063/