Update rbt patch for typing, the new Patch class, and crash fixes.
`rbt patch` now has type hints all throughout. This helped catch some
places where we were mixing types, leading to crashes. Notably, JSON
output would crash when dealing with conflicting files represented as
Unicode strings (which are expectd due to the type hints) rather than
byte strings.

It now uses the `Patch` class to track patches to apply, wrapped by a
small `PendingPatchInfo` dictionary with the rest of the information
that it needs to track.

This is in preparation for using new patching logic for more SCM-driven
patch application, which is in development now.

As a note, the `Patch` command has been renamed to `PatchCommand`, to
avoid conflicts with the `Patch` class. This will require those using
the development tree to re-run `pip install -e .` to register the new
entrypoint.

Reviewed at https://reviews.reviewboard.org/r/13970/