Split up rbtools.api.resource.
The resource definitions for RBTools are quite lengthy. While the bulk
of the current implementation all kind of needs to stick together, we
can split out individual resource-specific subclasses.

The bulk of this is just splitting up code without making any other
cleanup or functional changes. The only exception is DraftDiffResource,
which was created for a mimetype that doesn't exist in Review Board (the
server-side DraftDiffResource reports its mimetype as `diff` instead of
`draft-diff`. This resource subclass has been removed and the old name
is aliased to DiffResource.

Testing Done:
Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/14241/