Fix type of 'p' argument in SVN patching implementation.
The SVN implementation for `apply_patch` was trying to do an integer
comparison on the `--px` value, which was typed as a string. This change
fixes it so when a p-number is user supplied, we convert it to an int as
necessary.

Testing Done:
Ran unit tests (including fixed unit test for the `p` argument).

Reviewed at https://reviews.reviewboard.org/r/14217/