Pass the RBTools configuration to the HTTP/API layer.
We currently pass a lot of options all the way from `RBClient` to the
transport to `ReviewBoardServer`, and this list has continued to grow
over time. We will soon be needing another option, and instead of
continuing to add dedicated options, we're now passing the loaded
configuration instead.

`RBToolsConfig` was added in RBTools 5, and consolidates all of our
loaded options. This is easier to populate/load and pass around.
`RBClient`, `SyncTransport`, and `ReviewBoardServer` now all take this
as arguments.

If this is not provided, `ReviewBoardServer` will load a configuration.
This avoids breaking API compatibility with any custom scripts or hooks.
The loaded configuration will always be explicitly provided when being
run from a command.

Testing Done:
Unit tests passed.

Successfully posted this change for review.

Made use of the configuration in an in-progress change.

Reviewed at https://reviews.reviewboard.org/r/14221/