Add new resources, part 3/7.
This change adds the following new resource implementations:
- RepositoryItemResource
- RepositoryListResource
- RepositoryBranchItemResource
- RepositoryBranchListResource
- RepositoryCommitItemResource
- RepositoryCommitListResource
- RepositoryGroupItemResource
- RepositoryGroupListResource
- RepositoryInfoResource
- RepositoryUserItemResource
- RepositoryUserListResource

This involves one update to our base ListResource implementation (and
various related plumbing) to allow subclasses to force a specific class
type to be used for items. In the case of repository branches and
commits, these are list resources that do not have an associated item
endpoint or mimetype, and so the basic `_wrap_field()` and create
functionality was returning a plain `ItemResource`. For this case,
`ListResource` subclasses can now set `_item_resource_type` to force a
specific item resource subclass instead.

Testing Done:
- Ran unit tests.
- Used the new resources from a test script and saw that everything
  worked as expected.

Reviewed at https://reviews.reviewboard.org/r/14342/