Add new resources, part 7/7.
This change adds the following new resource implementations:
- ArchivedReviewRequestItemResource
- ArchivedReviewRequestListResource
- MutedReviewRequestItemResource
- MutedReviewRequestListResource

These resources are special because they do not support GET operations,
and the regular links introspection therefore doesn't work. The
implementation of these is therefore done as a custom layer that allows
us to get a "bare" instance of the resource, which can then be used for
`create` or `delete` as appropriate.

Testing Done:
- Ran unit tests.
- Used the new resources from a test script and saw that everything
  worked as expected.

Reviewed at https://reviews.reviewboard.org/r/14380/