Fix `rbt list-repo-types` when some repo types don't have dependencies.
If a dependency is missing, `rbt list-repo-types` would crash trying to
set it up. This change makes it so we catch `SCMClientDependencyError`
and just continue with the other tools.

Testing Done:
Was able to use `rbt list-repo-types` on a system that was missing some
things.

Reviewed at https://reviews.reviewboard.org/r/14321/