Add powerpack-based backend to ClearCase server_tool_names.
We had the powerpack backend for ClearCase listed in the `server_tool_ids`
field, but not `server_tool_names`. This meant that when running against an
older version of Review Board that does not allow filtering by ID,
RBTools would fail to find configured repositories. This change adds
that to the names field as well so we'll find and match all
ClearCase-based repositories, both with the old open-source and newer
powerpack backends.

Testing Done:
Ran a test with an older Review Board server and saw that we were
passing the correct argument to ?tool= in the API call.

Reviewed at https://reviews.reviewboard.org/r/14325/