Fix issues when running rbt commands outside of a repository with TREES.
When running some rbt commands outside of a repository (such as
`rbt api-get` or `rbt patch --print`) and using the `TREES`
configuration to define the server URL, rbtools would exit with a
message about being unable to find a repository. This was a regression
caused by the previous fix in 5.2 to allow these commands to use TREES.

This fix plumbs through the `tool_required` flag to
`scan_usable_client`, making it so we don't exit if we don't find a tool
at all.

Testing Done:
- Ran unit tests.
- Was able to successfully use `rbt patch --print`, `rbt patch --write`,
  and `rbt api-get` outside of any source tree when using `TREES` to
  define the server URL.
- Verified that `TREES` still worked as expected.

Reviewed at https://reviews.reviewboard.org/r/14396/