Avoid calling temp_data.update() in the Sphinx extension.
In Sphinx, we are looking to change `temp_data` from a dict to a
better-typed object (c.f. sphinx-doc/sphinx#13151). This change will be
backward compatible, but this repo is one of only two instances I have
found of `temp_data.update()`, and currently temp_data is not really
public API de jure. This change maintains current behaviour, and is
forward-compatible should we look to deprecate `.update()` in the
future.

Testing Done:
Built docs.