Add get_current_bookmark definition to base SCMClient.
The `get_current_bookmark` method was defined in the Mercurial and
Jujutsu backends, but there was no definition for it on the base
SCMClient. This change fixes that.

Testing Done:
No longer saw an ugly red underline under the usage of this method in
the post command implementation.

Reviewed at https://reviews.reviewboard.org/r/14472/