Remove deprecated warning from SCMClient.has_dependencies.
The `has_dependencies()` method has for a while had an `expect_checked`
parameter which will skip setup in the case where we know that we've
already done it. This raised a deprecation warning about being removed
in RBTools 5, but we never did so. This change modifies the warning to
be an exception instead, and updates all the relevant test cases to
match.

Testing Done:
Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/14487/