Fix posting Jujutsu changes that include deleted files.
The Jujutsu backend for RBTools gets the Git diff and then processes it
in order to turn short blob hashes into full hashes. This was broken for
deleted files in two ways:

- The regex that parsed index lines didn't catch deleted lines, because
  they didn't have any mode information.
- Trying to turn '00000000' into a full hash would fail because that
  hash doesn't actually exist in the repo.

This change fixes those two issues and adds a test to verify it.

This also fixes a test for an updated message in jj.

Testing Done:
- Ran unit tests.
- Posted a change that included a deleted file.

Reviewed at https://reviews.reviewboard.org/r/14563/